+++ pkg/client/start.go
@@ -1,7 +1,6 @@
 package client
 
 import (
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"os"
@@ -198,50 +197,4 @@ func (c *Client) checkPlex() bool {
 	}
 
 	return err != nil
-}
-
-// Temporary code? sure is ugly and should probably go in notifiarr.
-func (c *Client) testSnaps(url string) {
-	snaps, errs, debug := c.Config.Snapshot.GetSnapshot()
-	for _, err := range errs {
-		if err != nil {
-			c.Errorf("%v", err)
-		}
-	}
-
-	for _, err := range debug {
-		if err != nil {
-			c.Errorf("%v", err)
-		}
-	}
-
-	var (
-		plex *plex.Sessions
-		err  error
-	)
-
-	if c.Config.Plex != nil {
-		if plex, err = c.Config.Plex.GetSessions(); err != nil {
-			c.Errorf("%v", err)
-		}
-	}
-
-	b, _ := json.Marshal(&notifiarr.Payload{
-		Snap: snaps,
-		Plex: plex,
-	})
-
-	c.Printf("Snapshot Data:\n%s", string(b))
-
-	if url == "" {
-		return
-	}
-
-	if body, err := c.notifiarr.SendJSON(url, b); err != nil {
-		c.Errorf("POSTING: %v: %s", err, string(body))
-	} else if fields := strings.Split(string(body), `"`); len(fields) > 3 { //nolint:gomnd
-		c.Printf("Sent Test Snapshot to %s, reply: %s", url, fields[3])
-	} else {
-		c.Printf("Sent Test Snapshot to %s, reply: %s", url, string(body))
-	}
 }
